+++ src/test/run.ts
@@ -4,6 +4,7 @@ import "./sync.test";
 import "./query-demo.test";
 import "./api.test";
 import "./protocol.test";
+import "./exclusiveTask.test";
 import { run } from "./testKit";
 
 await run();